+++ changelog.md
@@ -34,6 +34,7 @@ Codion Change Log
 - fetchDepth renamed referenceDepth.
 - ForeignKeyDefinition.Builder.referenceDepth() added, ForeignKeyDefiner.foreignKey(referenceDepth) removed.
 - TransientAttribute.modifiesEntity() and Builder.modifiesEntity() renamed modifies().
+- SelectQuery renamed EntitySelectQuery.
 ### is.codion.framework.db.local
 - SelectQueriesTest improved.
 ### is.codion.framework.model